third_party/WebKit/LayoutTests/http/tests/inspector/console-xhr-logging-expected.txt

CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE WARNING: Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE MESSAGE: line 10: sending a %s request to %s
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled.

console-xhr-logging.html:5 sending a GET request to resources/xhr-exists.html
network-test.js:33 Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
network-test.js:34 Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
network-test.js:37 XHR finished loading: GET "http://127.0.0.1:8000/inspector/resources/xhr-exists.html".makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:12(anonymous function) @ VM:1
console-xhr-logging.html:5 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:37 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:14(anonymous function) @ VM:1
network-test.js:37 XHR finished loading: GET "http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html".makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:14(anonymous function) @ VM:1
console-xhr-logging.html:5 sending a POST request to resources/post-target.cgi
network-test.js:37 XHR finished loading: POST "http://127.0.0.1:8000/inspector/resources/post-target.cgi".makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:16(anonymous function) @ VM:1
console-xhr-logging.html:5 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
network-test.js:37 XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:18(anonymous function) @ VM:1
network-test.js:37 XHR failed loading: GET "http://localhost:8000/inspector/resources/xhr-exists.html".makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:18(anonymous function) @ VM:1
console-xhr-logging.html:5 sending a GET request to resources/xhr-exists.html
console-xhr-logging.html:5 sending a GET request to resources/xhr-does-not-exist.html
network-test.js:37 GET http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html 404 (Not Found)makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:14(anonymous function) @ VM:1
console-xhr-logging.html:5 sending a POST request to resources/post-target.cgi
console-xhr-logging.html:5 sending a GET request to http://localhost:8000/inspector/resources/xhr-exists.html
network-test.js:37 XMLHttpRequest cannot load http://localhost:8000/inspector/resources/xhr-exists.html. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.makeXHR @ network-test.js:37makeSimpleXHRWithPayload @ network-test.js:16makeSimpleXHR @ network-test.js:11requestHelper @ console-xhr-logging.html:6makeRequests @ console-xhr-logging.html:18(anonymous function) @ VM:1
